Rewrites cow milking & fixes grown-up Saigas being aggressive (#964)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This fixes milking(you had to cancel it due to an incorrect ! in front
of a do_after) as well as a problem where mobs, when growing up into
their adult forms, did not properly transfer their tamed status. i. e.
adult saigas grown from young ones were always aggressive.

This is a reimplementation of #941