Seelie(pixie) PR, part 2! (#968)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The little creatures of mischief and magic are back, and better than
ever!
I wont waste time re-hashing old features, for that please refer to the
original PR here:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/882

Instead, Ill be going over fun additions, some sad removals, and balance
changes.

Up first, male Seelies!
Male Seelies have been added, and are fully supported with custom erp
interactions and verbs.
However, please keep in mind that lore wise, Seelie are unable to breed
naturally, and are instead forces of nature that come into existence
through concentrations of *'magic'!*

We've also added the ability to select Bee Wings instead of the usual
Seelie Wings, should you wish to do so!

Now for the spells, some of the less impactful and outright buggy spells
have been removed in favor of new ones.
* Strip spell has been replaced with the ability to cast a weak water
bolt, that will extinguish targets on fire, water crops, and also very
briefly slow your target should they be living.
* Drain spell, which didn't work too well to begin with, has been
replaced with a very weak Regenerative Kiss! This will inject 1 unit of
red and 1 unit of blue potion into the target - but be careful, too many
kisses may be toxic! (By this I mean it also causes nausea buildup). 1
cast wont cause much buildup, but 2 casts and the target may start
vomiting!
* Replenish Nature spell, which could be used to refill completely empty
berry bushes, has been replaced with a host of new features. The direct
spell that replaced it is Tame Rous, which will do as the name suggests.

Tame Rous is greatly supported by Rous milking and riding! Yeehaw little
buddy!

![image](https://github.com/user-attachments/assets/2d0b4479-59b9-4112-ac1d-40123595bfcc)
![image](https://github.com/user-attachments/assets/f600605b-0821-47fe-b63b-f81fcf86f9e6)

![image](https://github.com/user-attachments/assets/89c7098c-af88-40b6-8988-587bee59682c)
To ride a Rous, simply mount it as you typically would a Saiga, and
you're off to the races!
*(Note these are a little faster than Saiga, and may need balancing
after testing)*

This is meant to help with animal husbandry, and make milk and meat a
little more common and plentiful.

For some fun interactions, Seelie can also now "Buckle" to a person,
riding along on their shoulder!
To do this, change to WEAK intent, and drag from the Seelie to the
Humenoid. (Both the Humenoid and Seelie can do the actual drag->drop
interaction)
To dismount, the Seelie can either press the RESIST key, or the person
being ridden can GRAB and DROP the Seelie.

![image](https://github.com/user-attachments/assets/deec86e9-0728-417f-9576-38a8a2442efa)
![image](https://github.com/user-attachments/assets/92c8363e-331d-4894-a00d-b124c6bedd68)

Furthermore, the Seelie player can now also choose whether to glow or
not, through a new Seelie Tab, similar to how church members access
devotion mechanics.
They can also use this tab to access a toggle for the previously
esoteric "Blessing / Curse" system. No longer is it tied to mood, and
now the player can choose whether to bless someone with +2 fortune, or
take it away and give them -2 instead.

Finally, there have been many bug fixes and balance changes across the
board.
* 6 slot custom job lowered to 5, and Seelie can no longer spawn as
Hostage.
* Furthermore, *only one single* Seelie player may spawn as Bandit. No
more than that ever.
* Lastly, the curse for wing ripping has been fixed, and will make
people deaf for a short while, apply confusion to randomize their
movement, and also give them a special luck debuff. Again: **This is all
temporary!!**
* Seelie luck lowered to only +2, to help prevent the ridiculous amount
of crits that were happening during last test merge

*Note on combat: After thorough testing, it has been found that Seelie
require on average 40 or more stabs with knives and blades to kill
someone. Between that and a flat 70% damage reduction for bows and
crossbows, we find this sufficient to not require further restrictions*

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Good for the game because it re-adds Seelies! Greatly missed, loved by
many, and loved to be hated by others.
(For a more serious answer check the previous PR linked above)

**MASSIVE SHOUTOUT to Thefenix323, just as last time!**
~ *I literally could not have done this without you.* ~
And a special thank you to all who helped make this possible and better
than ever!

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: TheFenix323 <thefenix323@gmail.com>
Co-authored-by: TheFenix323 <64550272+TheFenix323@users.noreply.github.com>
Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>