Allows you to sever external body parts, adds special biting effects during oral, and allows chained players to do some sex actions. (#796)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

CHANGELOG
- Implements a new "Sever external organs" surgery
- A surgery that requires far less tools and steps than normal organ
removal but leaves the patient heavily wounded on both a success or a
failure.
- To perform, use any tool that can make an incision to make one and
then any tool that can remove a limb.
- Works on eyes, tongues, ears, genitals, and any other part that is
externally visible.
- On a failure, leaves a heavy cut, on a success leaves a heavy cut and
a severed vein along with removing the desired organ.
- Intended to be an option for if you have healers on hand but not
surgical tools, or don't care if the "patient" bleeds out.
- Adds special handling for biting when performing oral sex acts.
- When performing some oral sex acts, bites will be redirected to the
appropriate area (namely groin and chest).
  - Biting can be done while chained when performing oral sex acts.
  - Biting gains a damage bonus during oral sex acts.
- Allows some sex acts to be done while chained.
- This includes most things that don't strictly need your hands, like
doing oral or riding.

Bugs noticed during testing that aren't a side effect of this from what
I know and that I haven't fixed yet:
- Leg chains do literally nothing.
  - It was never implemented by the looks of it.
- Unchained players in the stockade can do anything but move, including
using their arms and sex acts.

Otherwise, this has been tested as much as I reasonably can on my own
and should be relatively bug free. Do let me know if there are any
changes requested for this.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Requested by a friend who wishes to stay anonymous: "uh players being
r*ped constantly threaten to bite back but are literally unable to. the
need to do field surgery is constant, be it collecting eyes or punishing
criminals, but a doctor isn't always present."

For the surgery changes, it adds an element of realism and adds an
easier option for removing body parts with tools players will commonly
have on them while having a cost so as to help keep it from being
overpowered or even a good idea when proper tools and skills are
available.

For the biting changes, it adds an element of realism and risk to some
of the sex actions and allows people roleplaying defiance to have a way
to follow up on common threats.

For the sex acts while chained up changes, it allows players who are
chained to perform certain acts that they would reasonably be able to at
least somewhat do with just their hands bound and potentially gives
options for fighting back while bound in conjunction with the biting
changes by playing along and getting captors into vulnerable positions.

---------

Co-authored-by: GuppyLaxx <25938722+GuppyLaxx@users.noreply.github.com>