Defiants can no longer be sexually assaulted (#942)
RE-WRITE :

This PR makes it so Defiants cannot be sexually assault, or be sexually
assaulted. Defiants who are in combat mode cannot have the ERP menu
opened on them. Neither if they are sleeping, or if they have no
Client(disconnected)
It also makes it so you can always see if someone is defiant regardless
of range or your own defiant status.

# THIS PR DOES NOT AFFECT NON-DEFIANTS AT ALL. NON-DEFIANTS ARE ENTIRELY
UNPROTECTED MECHANICALLY TO ANY SEXUAL STUFF.


I tried to change as little as possible as to not mess with the current
systems fundamentally, in-case we or anyone wants to change this.