Slight weapon and buckler changes (#998)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

List of changes: 

1. Charging a flail smash now drains energy ro keep charged. With
standard 10 END you should be able to keep a flail charged for ~10
seconds.
2. Combat knife is now able to stab.
3. Combat knife is now called combat knife instead of just knife.
4. Combat knife is now called combat knife on the smithing list instead
of hunting knife.
5. Rapier cut now only deals 80% damage instead of 50% damage.
6. Buckler now properly uses armorsmithing when being crafted.
7. Buckler now needs 1 steel bar instead of 2.
8. Buckler now needs higher skill to craft since it is now cheaper on
material.
9. Combat knife now needs 1 steel bar instead of 2.
10. Combat knife now needs higher skill to craft since it is now cheaper
on material.
11. Added a missing / on the buckler code.
## Why It's Good For The Game

1. This is mostly prevent people from charging the flail smash for
unlimited time. Should prevent a bit smash spamming (hopefully) now that
it actually consumes stamina. Shouldn't change much for anyone that
doesn't charge the hit more than they should.
2. It's a combat knife, of course it should be able to stab. It's based
ot the "THIS is a knife" meme, and the only reason it didn't have a stab
option (i think, correct me if i'm wrong) is because it's based off the
cleaver code-wise, and a cleaver cannot stab.
3. Self explanatory.
4. Self explanatory.
5. It's a rapier, it's got an edge throughout most of the blade, it
should still be able to cut decently, not as good as a regular sword,
but it should still do some damage. Now it should deal 14 damage at
standard strength (slightly less than a dagger's cut)
6. All other shields use armorsmithing, why wouldn't the buckler? Known
issue, now fixed.
7. You mean to tell me that the buckler shield needs the same amount of
material as a tower shield? It's a small shield used for skill over flat
defense, it's already underused as it is, no need to make it
unreasonably expensive on material.
8. Self explanatory.
9. Why would a knife half the size of a sword take twice as much steel
to make? The weapon doesn't even get used that often probably because of
the high cost and low damage.
10. Self explanatory.
11. This probably does nothing, but i just HAD to fix it. It haunted me.
And i _hated_ it.

---------

Co-authored-by: GuppyLaxx <25938722+GuppyLaxx@users.noreply.github.com>