PSYDON added to roles (#1009)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

ALL_DIVINE_PATRONS is applied to Knight, captain, gatemaster and bog
master. Now changed to ALL_NON_INHUMEN_PATRONS which just had psydon
added.

## Why It's Good For The Game

I think its cool having more roles be able to be psydonites. The king
can literally be a psydonite but for some reason a good chunk of his
most important retinue cannot. Not to mention the bog and gate masters
as well.