Changes the satchel slots to 4 (#976)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Changes the satchels slots to 5, and it's maximum weight to 14 (Max 4
normal items and 1 small)

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
The satchel (currently) has 3 slots and 21 max weight, it's clearly
intended to be a more versatile and lighter version of the backpack, but
it usually fails at that since it's low capacity. It even has a max
weight of 21, which is 7 normal items max.
This hopefully will make carrying things from one place to the other
less of a trouble without a backpack, and help with making carrying more
than three things slightly less annoying.

---------

Co-authored-by: GuppyLaxx <25938722+GuppyLaxx@users.noreply.github.com>