Makes Rockhill Great Again! (#999)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## Adds "dense rock" to funnel players to utilize the lower miners in
the bog, or tunnel closer to the Bog Gate.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Players have begun to make pickaxes and other crude items to bypass the
Bog-Gate entirely by simply building around it. This alleviates the
issue by forcing players to either confront the Bog-Guard when
attempting to bypass the gate, or use a more undetectable method.

Effectively just makes it so when players want to bypass the Gate, they
will do so in an obvious manner through mining directly to the left of
the Bog Gate, mining to the rooftops of the Bog Gate house, or actually
take effort as they need to utilize the bog mines that are under the bog
to completely bypass the Gate undetected.

Additionally pulls PR #997