New attempt at removing Bandit Seelies (#1012)
Now it ask the client what race they are selecting in the lobby.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds new logic to check if the player is chosing to be a Seelie wile
rolling for bandit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes Seelies being able to spawn as Bandits.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>