[TM Ready] Lupian language update (#803)
Adds a language for both Lupians and Vulps including a new font for
them.

Lupians also get a new accent that makes them roll their r sounds and
makes most of their swear words czech. Vulps do not get the accent.

This accent doesn't have that much affect in game, most of the
translations only trigger when swearing with an exclamation at the end,
you can still say cunt and fuck and shit without it changing.

Lore slop, variety, now people can stop complaining elves and orcs are
the only ones with accents.
The font used is Emperialisme, which is freeware
https://www.fontspace.com/emperialisme-font-f102330

![image](https://github.com/user-attachments/assets/8d817599-ba87-44dc-b569-0860dd5601c2)

---------

Co-authored-by: zoktiik <50035052+zoktiik@users.noreply.github.com>