[FIX] Sticks - dropping from hands on pickup (#1041)
- deletes the stick in your hand before adding the new one, to prevent
dropping while picking up sticks.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

This is a simple fix that prevents sticks from dropping on the ground
when collecting with one hand full.

According to the documentation for "put_in_hands" the item will drop if
there is no available hand. This change simply deletes the sticks in the
user's hand before the new pile is added.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Makes picking up sticks NOT cancer. 

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Really speaks for itself.