Oo yeah woo yeh oo yeah horns baby
Makes horn more useful on a multiple levelled map. It now returns the area if you're in a notable landmark. So, the town, azure coast, basin, grove, etc. Horn-hearers will now be able to trust the final line of information, since being on a different z busts the direction/distance.

Later I will figure if that problem can be solved, but for now, the Hornening shall be a useful tool if you're in any area. Also, makes them a bit more accurate at shorter ranges and gives Wardens back the old bogguard horn sound.

Ex) "I hear the horn of the Wardens within 20 meters, somewhere to the northwest, in the Church!"

"I hear the horn of the Wardens very far, to the southwest, in the Azure Grove!"