The PR Giveth: Added actions for penetrating with tails (#1062)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds two new interactions using the tail. These can be used even if the
giver is fully clothed.
Only works on kobold and lizard tails.
Seelies get destroyed by it, as one would expect.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Tails are pretty phallic, and this translates nicely into the mechanical
sex we've got.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->