The PR Taketh Away: Allows tails, horns, and frills to be removed in character customization (#1061)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes it so your little freaks of nature can be missing their tail.
Works on all races.
Same thing for horns and frills, for Sisseans, Drakians, and Kobolds.
Also has the side-effect of allowing kobolds to pick a color for their
tail.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Some characters might not have tails. Some might have lost them.
Prevents the rest of players from having a delicious meal of lizard
tail.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>