Metal uniformity V2- Red Iron Edition (#1055)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Oh, boy, this is a big one!

This applies a new iron palette to most (if not all, I might have missed
some) iron items in the game. This also applies the steel palette to
armour, and a new gold palette to just about everything that used the
old one. Mammon included. It's difficult to include in a single
screenshot.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
![Screenshot 2024-09-26
195708](https://github.com/user-attachments/assets/f91eedc1-80e5-4c5c-850f-888d1fea7e72)
![Screenshot 2024-09-26
195729](https://github.com/user-attachments/assets/d223e4d9-2d2e-46b0-af03-1929e3689e6a)
![Screenshot 2024-09-26
195910](https://github.com/user-attachments/assets/a886d90a-07f3-4ff8-bc88-79669df5bb9f)
![Screenshot 2024-09-26
195930](https://github.com/user-attachments/assets/33a22ee8-8075-4068-a5d6-ac50056dc6c9)
![Screenshot 2024-09-26
200357](https://github.com/user-attachments/assets/c29248db-b353-4e3f-89f6-c3a3a5cd5752)



## Why It's Good For The Game
The old iron palette was pretty ugly. Armour did not follow the steel
palette consistently. Now both of these are fixed.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>