Armory Reinforcement (#1080)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Due to individuals being incapable of tact, the Armories throughout
rockhill and their associated barracks' are receiving dire
reinforcement.

Stone windows are replacing the old creaky wooden windows throughout the
Bog Barracks and Watchman Head Quarters.

## Why It's Good For The Game

It isn't good or bad for the game. It should fixes the problem of those
incapable of actually role playing of rushing the Armories for materials
10 minutes into a round.