Trays no longer spill open containers' reagents (#1007) (#1068)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes #1007.
Trays will uniquely refuse to spill any liquids from "`spillable`"
reagent_containers, such as cups and bottles, that are put on them. For
reference, liquid "spilling" (deleting 3 units of each present reagent)
is generally supposed to occur:
- when putting an open r-container into a bag
- when grabbing a bag with open r-containers,
- when moving around with a bag that has open r-containers

Trays will now ignore all 3 of these cases. If you want to make certain
things like throwing a tray cause it to drop & spill r-containers, be my
guest. But that's out-of-scope for this PR.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Using trays for carrying around open beverage containers both makes
sense, and is good for roleplay. The tray becomes substantially more
interesting of a storage container with this PR.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>