Lfisher bug fix (#1072)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Legendary fisher was spawning with no backpack since it was putting the
backpack on the right side, when they only fit on the left side, this
fixes it.

## Why It's Good For The Game

Less bugs > Same amount of bugs
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->