Craftable arming jackets + fem sprite for it (#1030)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Arming jackets now have a sprite for females, because right now they
just don't.
Arming jackets can now be crafted. Identical cost to a gambeson, since
it is completely identical to a gambeson stat/coverage wise.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Women not having sprites for a clothing they can wear makes them appear
nude. That's bad.
Craftable arming jackets because why not, They are currently only
available from from spawning as the Hand or Marshal, despite not
actually being any special. Can add a silk cost to the recipe if we want
to keep it as a "higher end" piece.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->