No more legendary pilgrims (#1091)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
changes 	
        maximum_possible_slots = 1
        and
	pickprob = 0.0001
as per maintainer request 

![image](https://github.com/user-attachments/assets/147d099a-c124-4a34-a7dc-dc6a75db633a)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
"legendary" my ass
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->