AI NPC FIX - Dungeon Revert (#1042)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I don't know exactly why, but the entire PR breaks the AI completely. I
tried ONLY reverting the zmanager changes and the fix was partial. So
the entire PR must be reverted and whoever wants to add the Dungeon back
must first figure out whats wrong with it.

BEFORE:

https://github.com/user-attachments/assets/7b70e4ba-f711-49fa-8d75-508a5205b72e

AFTER:

https://github.com/user-attachments/assets/56fb447e-3002-4049-a8fb-6396e219c3c6

The PR in question:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/760

---------

Co-authored-by: Gup <guppylaxx@gmail.com>