Taskbar icon update. (#1109)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes the icon on the taskbar to a cute little rat!

![image](https://github.com/user-attachments/assets/dea944b7-78a2-4618-8c40-3a8afb97d47c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Removing Blackstone fragment.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>