Allows shunned races in the position of Head Butler/Maid. (#1056)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Same as https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1036, but I
had to recreate my branch due to Github breaking. Stated as in title.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Just pulling my description from the last attempt:
"If shunned races are allowed to be Hand, why should not a position of
ultimate servitude not be fit for them? This (aside from Hand) would be
representation of the highest they might ever climb and the closest to
respect they would ever be afforded."
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>