Limb Artery Severity Fix (#1102)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Halves the lethality of limb arteries, giving you double the time to fix
this wound before it becomes fatal, if you start suturing the wound it
should reduce bleeding even more too.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes combat last bit longer, neck and heart arteries are still super
lethal just made the easier to hit arteries less lethal for the sake of
not dying so fast in-game.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->