Terra Bog Changes Part 1 (#993)
Taken away:
Terrabog Smithy
Lava Mines

Added:
Ravine with new Lava Mines and abandoned mining village Abandoned tribal
camp overrun with wolves
Abandoned Totally not Zizo church with mini dungeon and loot Goblin Cave
Mini dungeon with loot

Wolves in a few locations on the map start to make terrabog slightly
scarier to travel.

Took out lights from the underground caves with only a few lights left.

Bog Smithy is taken out for now, and may be re added later in another
part of the updates. But there are two currently findable riddle of
steels in the mini dungeons as a compensation for bog smithies. But
require the dungeons to be conquered first.

With this first update, we will now give players more reason to explore
the terrabog. As well as give bandits more people to rob throughout the
round instead of just relying on going to attack the town or boggate.

As well as three mini dungeons that allow for players to not only loot,
but conquer them in the terrabog and allow them to rebuild with some
effort if they wish to set up small groups.

The goal with these updates to the terrabog. Is to slowly reshape it,
making it more dangerous, but also more rewarding to explore. As well as
giving antags more reason to stick around the area instead of round
rushing keep or town.

Abandoned Tribal Camp:

https://cdn.discordapp.com/attachments/1240810578029576292/1284791342618247250/Screenshot_2024-09-15_163423.png?ex=66e7eaae&is=66e6992e&hm=b7959fdbc272e5ba810706cc1bada6e9228889d8aec05798fdaebc29fe3a135b&

Abandoned Totally not Zizo Church

https://cdn.discordapp.com/attachments/1240810578029576292/1284791343062847599/Screenshot_2024-09-15_163435.png?ex=66e7eaae&is=66e6992e&hm=7a861f78ee242b932d53f8bee2017062c96a5a4e60b9690155c1b84f21321fa7&

https://cdn.discordapp.com/attachments/1240810578029576292/1284791343419228260/Screenshot_2024-09-15_163440.png?ex=66e7eaae&is=66e6992e&hm=e294a8656d3caa9fce96f43316d84d2000a313dcdad8d8a76be558cc1fb856d5&

https://cdn.discordapp.com/attachments/1240810578029576292/1284791343666827335/Screenshot_2024-09-15_163447.png?ex=66e7eaae&is=66e6992e&hm=6ccdd91240e3de6ab7089f658ef474926c2e51fc402ee02bf919cc496bf5bcb1&

Gobbo Cave

https://cdn.discordapp.com/attachments/1240810578029576292/1284791343939584050/Screenshot_2024-09-15_163501.png?ex=66e7eaae&is=66e6992e&hm=a7ecad445718f6a47ce7daf83032d2bb1bd376abb2e8714df74b0abb700cf65a&

Ravine / abandoned mining camp

https://cdn.discordapp.com/attachments/1240810578029576292/1284791344191111208/Screenshot_2024-09-15_163512.png?ex=66e7eaae&is=66e6992e&hm=dab56da80ebd18d042bfb607f000662ba825e0a47977eb2722a97bd9efd45b44&

https://cdn.discordapp.com/attachments/1240810578029576292/1284791344463745065/Screenshot_2024-09-15_163520.png?ex=66e7eaae&is=66e6992e&hm=584f528445404e8de875b20f9510e27b495342f59f8bae00c7519139fe080c75&

https://cdn.discordapp.com/attachments/1240810578029576292/1284791344774250517/Screenshot_2024-09-15_163527.png?ex=66e7eaae&is=66e6992e&hm=0bd87e13f1cb46186ee0201cfe96b3d36f7a8d44c0cc513ef5eb6a6396739bca&

---------

Co-authored-by: Gup <guppylaxx@gmail.com>