Fablefield just golaird wave as final step (#1034)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes a final step of the fablefield wave thats just the Goliard.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
For the solo clowns in the audience.

EDIT: Lowered strength requirement for the repenta en since the
inquisitor is often too feeble to carry his own whip.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->